added authentication and authorization to deleting posts using api requests

+++ controllers/api/v1/posts_api_controller.js
@@ -24,18 +24,19 @@ module.exports.destroy = async function(req,res){
     try{
       let post = await Post.findById(req.params.id);
       // .id means we are converting the object id stored in ._id to string for comparison
-    //   if(post.user == req.user.id){
+      if(post.user == req.user.id){
        post.remove();
        await Comment.deleteMany({post : req.params.id /* or post.id */});
        
        return res.json(200,{
            message : "Post and associated Comments deleted !"
        });
-    //   }
-    //   else{
-    //     req.flash('error', 'You cannot delete this post !');
-    //     return res.redirect('back');
-    //   } 
+      }
+      else{
+        return res.json('401',{
+            message : "You cannot delete this post !"
+        });
+      } 
     } 
     catch(err){
       console.log(err);
